Get rid of `converse-uniview` plugin
Instead of overriding createChatBox, set 'hidden' flag via defaults
